libwasi: rename host dirent things
